// Project front/front-end/src/app/model/caee.model.ts
export interface CAEE {
    issuerUsername: string;
    subjectUsername: string;
    subjectOrganization: string;
    subjectOrganizationUnit: string;
    subjectCountry: string;
    subjectEmail: string;
    newPKIssuerPassword:string;
    newKeyStorePassword: string;
    startDate: Date;
    endDate: Date;
  }